Add test for try to access post intervention without token
